# Optimize varint encoding