This optimizes varint encoding by turning the loop into a for loop which causes the compiler to see that it can unroll the loop.

Bench:
```
varint/small/encode     time:   [61.577 ns 61.786 ns 62.016 ns]
                        change: [-18.793% -17.983% -17.024%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe

varint/medium/encode    time:   [312.28 ns 313.01 ns 313.78 ns]
                        change: [-13.850% -13.002% -12.209%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

varint/large/encode     time:   [533.79 ns 535.11 ns 536.43 ns]
                        change: [-26.542% -25.921% -25.275%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe

varint/mixed/encode     time:   [308.34 ns 309.11 ns 309.87 ns]
                        change: [-23.418% -22.569% -21.718%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
```