@LucioFranco 2022-12-12
I am not seeing the same improvements on my laptop 

```
varint/small/encode     time:   [254.51 ns 255.08 ns 255.78 ns]
                        change: [+4.9943% +5.3430% +5.6646%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 14 outliers among 100 measurements (14.00%)
  6 (6.00%) high mild
  8 (8.00%) high severe

varint/small/decode     time:   [230.34 ns 230.96 ns 231.62 ns]
                        change: [+0.0577% +0.4059% +0.7424%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

varint/small/encoded_len
                        time:   [57.304 ns 57.385 ns 57.477 ns]
                        change: [-0.3950% -0.1263% +0.1243%] (p = 0.35 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe

varint/medium/encode    time:   [1.2231 us 1.2262 us 1.2301 us]
                        change: [+5.3350% +5.6709% +6.0007%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

varint/medium/decode    time:   [240.14 ns 240.56 ns 241.05 ns]
                        change: [-0.8120% -0.5223% -0.2217%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

varint/medium/encoded_len
                        time:   [57.344 ns 57.443 ns 57.556 ns]
                        change: [-0.3119% -0.0467% +0.2367%] (p = 0.73 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe

varint/large/encode     time:   [2.3195 us 2.3250 us 2.3310 us]
                        change: [-1.1808% -0.8771% -0.5842%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild

varint/large/decode     time:   [352.08 ns 352.89 ns 353.81 ns]
                        change: [-0.3626% -0.0670% +0.2361%] (p = 0.66 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

varint/large/encoded_len
                        time:   [57.386 ns 57.487 ns 57.598 ns]
                        change: [-0.5971% -0.2721% +0.0337%] (p = 0.10 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild

varint/mixed/encode     time:   [1.4749 us 1.4777 us 1.4808 us]
                        change: [-2.2308% -1.9438% -1.6452%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

varint/mixed/decode     time:   [286.76 ns 287.35 ns 288.00 ns]
                        change: [-0.4404% -0.1303% +0.1851%] (p = 0.41 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  11 (11.00%) high mild
  1 (1.00%) high severe

varint/mixed/encoded_len
                        time:   [57.407 ns 57.537 ns 57.689 ns]
                        change: [-0.5136% -0.1899% +0.1612%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
```