@LucioFranco 2022-12-12
Could you explain more what system you ran those benchmarks on?